Move performance tests to test/e2e/scalability subdirectory
