+++ test/e2e/BUILD
@@ -25,7 +25,7 @@ go_test(
         "//test/e2e/extension:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/metrics:go_default_library",
-        "//test/e2e/perf:go_default_library",
+        "//test/e2e/scalability:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/storage:go_default_library",
         "//test/e2e/workload:go_default_library",
@@ -58,7 +58,6 @@ go_library(
         "dns_common.go",
         "dns_configmap.go",
         "e2e.go",
-        "empty.go",
         "empty_dir_wrapper.go",
         "etcd_failure.go",
         "events.go",
@@ -140,7 +139,6 @@ go_library(
         "//test/e2e/generated:go_default_library",
         "//test/e2e/manifest:go_default_library",
         "//test/e2e/metrics:go_default_library",
-        "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
         "//test/e2e/workload:go_default_library",
@@ -229,8 +227,8 @@ filegroup(
         "//test/e2e/generated:all-srcs",
         "//test/e2e/manifest:all-srcs",
         "//test/e2e/metrics:all-srcs",
-        "//test/e2e/perf:all-srcs",
         "//test/e2e/perftype:all-srcs",
+        "//test/e2e/scalability:all-srcs",
         "//test/e2e/scheduling:all-srcs",
         "//test/e2e/storage:all-srcs",
         "//test/e2e/testing-manifests:all-srcs",
