+++ test/e2e/e2e_test.go
@@ -24,7 +24,7 @@ import (
 	_ "k8s.io/kubernetes/test/e2e/cluster-logging"
 	_ "k8s.io/kubernetes/test/e2e/extension"
 	"k8s.io/kubernetes/test/e2e/framework"
-	_ "k8s.io/kubernetes/test/e2e/perf"
+	_ "k8s.io/kubernetes/test/e2e/scalability"
 	_ "k8s.io/kubernetes/test/e2e/scheduling"
 	_ "k8s.io/kubernetes/test/e2e/storage"
 	_ "k8s.io/kubernetes/test/e2e/workload"
