+++ test/e2e/scalability/empty.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package scalability
 
 import (
 	"time"
@@ -22,7 +22,6 @@ import (
 	"k8s.io/api/core/v1"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/test/e2e/framework"
-	"k8s.io/kubernetes/test/e2e/perf"
 
 	. "github.com/onsi/ginkgo"
 )
@@ -43,7 +42,7 @@ var _ = framework.KubeDescribe("Empty [Feature:Empty]", func() {
 	})
 
 	It("starts a pod", func() {
-		configs, _, _ := perf.GenerateConfigsForGroup([]*v1.Namespace{f.Namespace}, "empty-pod", 1, 1, framework.GetPauseImageName(f.ClientSet), []string{}, api.Kind("ReplicationController"), 0, 0)
+		configs, _, _ := GenerateConfigsForGroup([]*v1.Namespace{f.Namespace}, "empty-pod", 1, 1, framework.GetPauseImageName(f.ClientSet), []string{}, api.Kind("ReplicationController"), 0, 0)
 		if len(configs) != 1 {
 			framework.Failf("generateConfigs should have generated single config")
 		}
